修改elementId、offset为int64_t

+++ window_scene/session/host/src/zidl/session_stub.cpp
@@ -494,8 +494,8 @@ int SessionStub::HandleTransferAccessibilityEvent(MessageParcel& data, MessagePa
 {
     sptr<Accessibility::AccessibilityEventInfoParcel> infoPtr =
         data.ReadStrongParcelable<Accessibility::AccessibilityEventInfoParcel>();
-    int32_t uiExtensionIdLevel = 0;
-    if (!data.ReadInt32(uiExtensionIdLevel)) {
+    int64_t uiExtensionIdLevel = 0;
+    if (!data.ReadInt64(uiExtensionIdLevel)) {
         WLOGFE("read idVect error");
         return ERR_INVALID_DATA;
     }
